+++ src/components/Header/Header.hooks.ts
@@ -1,13 +1,6 @@
 import { useUser } from '../../store/users/store';
-import { UserState } from './Header.type';
-import create from 'zustand';
-import StoreState  from './Header.type';
+import { UserState } from './Header.types';
 
 export const useHeader = () => {
   return useUser((state: UserState) => state.isLoggedIn);
 };
-
-export const useStore = create<StoreState>((set) => ({
-  isLoginModalOpen: false,
-  toggleLoginModal: () => set((state) => ({ isLoginModalOpen: !state.isLoginModalOpen })),
-}));
