--- src/components/Header/Header.type.ts
@@ -1,10 +1,0 @@
-export interface HeaderProps {}
-
-export interface UserState {
-  isLoggedIn: boolean;
-}
-
-export default interface StoreState {
-  isLoginModalOpen: boolean;
-  toggleLoginModal: () => void;
-}
