+++ src/components/Header/Header.types.ts
@@ -1,0 +1,5 @@
+export interface HeaderProps {}
+
+export interface UserState {
+  isLoggedIn: boolean;
+}
