+++ src/components/LoginModal/LoginModal.hooks.ts
@@ -1,0 +1,7 @@
+import StoreState  from './LoginModal.types';
+import create from 'zustand';
+
+export const useStore = create<StoreState>((set) => ({
+  isLoginModalOpen: false,
+  toggleLoginModal: () => set((state) => ({ isLoginModalOpen: !state.isLoginModalOpen })),
+}));
